fix bug when it comes to small segments without neighbors

+++ ni_vision/src/ni/modules/ni/layers/mapareafilter.h
@@ -47,7 +47,7 @@ public:
     void Activate(const elm::Signal &signal);
 
 protected:
-    int getNeighbors(float vtx_id, const elm::GraphAttr &seg_graph, std::vector<Surface> &neighbors, elm::VecF &neigh_sizes) const;
+    int getNeighbors(int vtx_id, const elm::GraphAttr &seg_graph, std::vector<Surface> &neighbors, elm::VecF &neigh_sizes) const;
 
     /**
      * @brief get sizes of all segments from map
